[DDW-781] Blank Screen Fix - Default Selection.
