+++ source/main/utils/buildAppMenus.js
@@ -1,5 +1,5 @@
 // @flow
-import { app, globalShortcut, Menu, BrowserWindow } from 'electron';
+import { app, globalShortcut, Menu, BrowserWindow, dialog } from 'electron';
 import { get } from 'lodash';
 import { environment } from '../environment';
 import { winLinuxMenu } from '../menus/win-linux';
@@ -10,6 +10,7 @@ import { CardanoNode } from '../cardano/CardanoNode';
 import { DIALOGS, SCREENS } from '../../common/ipc/constants';
 import { showUiPartChannel } from '../ipc/control-ui-parts';
 import { getLocale } from './getLocale';
+import { getTranslation } from './getTranslation';
 
 const localesFillForm = {
   'en-US': 'English',
@@ -23,6 +24,21 @@ export const buildAppMenus = async (
 ) => {
   const { ADA_REDEMPTION } = SCREENS;
   const { ABOUT, BLOCK_CONSOLIDATION, DAEDALUS_DIAGNOSTICS } = DIALOGS;
+
+  const {
+    isMacOS,
+    version,
+    apiVersion,
+    network,
+    build,
+    installerVersion,
+    os,
+    buildNumber,
+    isWindows,
+    isInSafeMode,
+  } = environment;
+  const translations = require(`../locales/${locale}`);
+  const networkLocale = getLocale(network);
 
   const openAboutDialog = () => {
     if (mainWindow) showUiPartChannel.send(ABOUT, mainWindow);
@@ -54,20 +70,35 @@ export const buildAppMenus = async (
     safeExitWithCode(22);
   };
 
-  const {
-    isMacOS,
-    version,
-    apiVersion,
-    network,
-    build,
-    installerVersion,
-    os,
-    buildNumber,
-  } = environment;
-
-  const translations = require(`../locales/${locale}`);
-
-  const networkLocale = getLocale(network);
+  const toggleOnSafeMode = item => {
+    const translation = getTranslation(translations, 'menu');
+    const gpuSafeModeDialogOptions = {
+      buttons: [
+        translation('helpSupport.gpuSafeModeDialogConfirm'),
+        translation('helpSupport.gpuSafeModeDialogNo'),
+      ],
+      type: 'warning',
+      title: isInSafeMode
+        ? translation('helpSupport.gpuSafeModeDialogTitle')
+        : translation('helpSupport.nonGpuSafeModeDialogTitle'),
+      message: isInSafeMode
+        ? translation('helpSupport.gpuSafeModeDialogMessage')
+        : translation('helpSupport.nonGpuSafeModeDialogMessage'),
+      defaultId: isWindows ? 1 : 2,
+      cancelId: 2,
+      noLink: !!isMacOS,
+    };
+    dialog.showMessageBox(mainWindow, gpuSafeModeDialogOptions, buttonId => {
+      if (buttonId === 0) {
+        if (isInSafeMode) {
+          restartWithoutSafeMode();
+        } else {
+          restartInSafeMode();
+        }
+      }
+      item.checked = isInSafeMode;
+    });
+  };
 
   const supportRequestData = {
     frontendVersion: version,
@@ -86,8 +117,7 @@ export const buildAppMenus = async (
     openAboutDialog,
     openDaedalusDiagnosticsDialog,
     openAdaRedemptionScreen,
-    restartInSafeMode,
-    restartWithoutSafeMode,
+    toggleOnSafeMode,
     openBlockConsolidationStatusDialog,
   };
